Julian dev

+++ src/main/java/kernbeisser/CustomComponents/ShoppingTable/ShoppingCartModel.java
@@ -7,12 +7,12 @@ import java.util.Collection;
 import java.util.HashMap;
 
 public class ShoppingCartModel implements Model {
-    private final int userValue;
-    private final int userSurcharge;
+    private final double userValue;
+    private final double userSurcharge;
     private HashMap<ShoppingItem,ShoppingItem> shoppingItems = new HashMap<>();
 
 
-    ShoppingCartModel(int userValue, int userSurcharge) {
+    ShoppingCartModel(double userValue, double userSurcharge) {
         this.userValue = userValue;
         this.userSurcharge = userSurcharge;
     }
@@ -26,7 +26,7 @@ public class ShoppingCartModel implements Model {
         }
     }
 
-    public int getUserValue() {
+    public double getUserValue() {
         return userValue;
     }
 
@@ -34,7 +34,7 @@ public class ShoppingCartModel implements Model {
         return shoppingItems.values();
     }
 
-    public int getUserSurcharge() {
+    public double getUserSurcharge() {
         return userSurcharge;
     }
 }
